+++ react/components/comparisonList/ComparisonSummary.tsx
@@ -1,51 +1,34 @@
 import React, { useMemo } from 'react'
-import { path } from 'ramda'
-import productsQuery from 'vtex.store-resources/QueryProduct'
-import { useQuery } from 'react-apollo'
 import { ExtensionPoint } from 'vtex.render-runtime'
 import { mapCatalogProductToProductSummary } from '../utils/normalize'
+import styles from './comparisonList.css'
 
 interface Props {
+  product: Product
   productToCompare: ProductToCompare
+  columnStyles: Styles
 }
 
-const ComparisonSummary = ({ productToCompare }: Props) => {
-  const { data: productsResponse, error, loading } = useQuery(productsQuery, {
-    skip: !(productToCompare && productToCompare.productId),
-    variables: {
-      identifier: {
-        field: 'id',
-        value: productToCompare.productId,
-      },
-    },
-  })
-
+const ComparisonSummary = ({
+  product,
+  productToCompare,
+  columnStyles,
+}: Props) => {
   const productDetails = useMemo(() => {
-    const selectedProduct: Product | undefined = path(
-      ['product'],
-      productsResponse
-    )
-
-    return selectedProduct
-      ? mapCatalogProductToProductSummary(
-          selectedProduct,
-          productToCompare.skuId
-        )
+    return product
+      ? mapCatalogProductToProductSummary(product, productToCompare.skuId)
       : {}
-  }, [productToCompare, productsResponse])
+  }, [productToCompare, product])
 
-  return loading ? (
-    <div>Loading...</div>
-  ) : error ? (
-    <div>{error}</div>
-  ) : (
+  return (
     <div
       key={
         productToCompare && productToCompare.productId
           ? productToCompare.productId
           : 'product'
       }
-      className="w-20"
+      className={`${styles.comparisonProductCol} ma1 pa3`}
+      style={columnStyles}
     >
       <ExtensionPoint id="product-summary" product={productDetails} />
     </div>